Main screen UI rework.

+++ Assets/Scripts/Data/ChartJSON.cs
@@ -64,7 +64,7 @@ public class Page
 public class Tempo
 {
     public int tick;
-    public int value;
+    public long value;
     [NonSerialized, JsonIgnore]
     public double time;
 }
